+++ recursion/getRange/getRange.js
@@ -1,5 +1,5 @@
 // `getRange`
-// Write a JavaScript program to get the integers in range (x, y)
+// Write function to get the integers in range (x, y)
 const getRange = (start, end) =>
   end > start ? [...getRange(start, end - 1), end] : [start];
 export default getRange;
