+++ recursion/getRange/readme.md
@@ -1,0 +1,7 @@
+# getRange
+
+---
+Write recursive `getRange` function to get the integers in range (x, y)
+```js
+getRange(1, 6); // [1, 2, 3, 4, 5, 6]
+```
